Errors after moving functions resolved
