+++ commands/eval/constants.py
@@ -1,0 +1,5 @@
+FILE_NAMES = {
+    "versions": "versions.diff",
+    "scout_yaml": "scout_yaml.diff",
+    "qc": "qc.diff",
+}
